Refactor: Moved dummy codes 미사용 코드 옮김
